requirements.txt
asgiref==3.8.1
Django==5.2
greenlet==3.2.1
pillow==11.2.1
playwright==1.51.0
pyee==12.1.1
sqlparse==0.5.3
typing_extensions==4.13.2
